dom/network/tests/test_networkstats_enabled_perm.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test to ensure NetworkStats is not accessible when it is disabled</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

"use strict";

SimpleTest.waitForExplicitFinish();

// Test to ensure NetworkStats is not accessible when it is disabled
SpecialPowers.addPermission("networkstats-manage", true, document);
SpecialPowers.pushPrefEnv({'set': [["dom.mozNetworkStats.enabled", true]]},
                          function(){
  ok(SpecialPowers.hasPermission("networkstats-manage", document),
     "Has permission 'networkstats-manage'.");

  ok(SpecialPowers.getBoolPref("dom.mozNetworkStats.enabled"),
     "Preference 'dom.mozNetworkStats.enabled' is true.");

  ok('mozNetworkStats' in navigator, "navigator.mozNetworkStats should exist");

  ok(navigator.mozNetworkStats instanceof SpecialPowers.Ci.nsIDOMMozNetworkStatsManager,
     "navigator.mozNetworkStats should be a nsIDOMMozNetworkStatsManager object");

  SpecialPowers.removePermission("networkstats-manage", document);
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
